Move IP-Adapter test_total_downscale_factor(...) to pipeline test file (requested in code review).
